src/requirements.txt
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
protobuf==3.20.0
python-dotenv==0.20.0
Werkzeug==2.1.1
zipp==3.8.0
